+++ test/dummy/app/controllers/application_controller.rb
@@ -5,6 +5,8 @@ class ApplicationController < ActionController::Base
   # Changes to the importmap will invalidate the etag for HTML responses
   stale_when_importmap_changes
 
+  helper RecordingStudio::LayoutHelper if defined?(RecordingStudio::LayoutHelper)
+
   layout :application_layout
 
   before_action :authenticate_user!
@@ -13,7 +15,7 @@ class ApplicationController < ActionController::Base
   private
 
   def application_layout
-    devise_controller? ? "application" : "flat_pack_sidebar"
+    devise_controller? ? "application" : "recording_studio/default_layout"
   end
 
   def set_current_actor